requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
drf-spectacular==0.18.1
httpie==2.4.0
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
jsonschema==3.2.0
Markdown==3.3.4
Pillow==8.3.2
Pygments==2.10.0
pyrsistent==0.18.0
PySocks==1.7.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0
